fix: restructure to replicate hiro api endpoints
