+++ functions/hiro-api/v2/info.ts
@@ -1,5 +1,5 @@
 import { PagesFunction } from "@cloudflare/workers-types";
-import { CloudflareBindings } from "../../worker-configuration";
+import { CloudflareBindings } from "../../../worker-configuration";
 
 export const onRequest: PagesFunction<CloudflareBindings> = async (context) => {
   const id = context.env.HIRO_API.idFromName("hiro-api");
